Rename format 4 to format 2
